+++ jgiven-junit/src/test/java/com/tngtech/jgiven/junit/DataProviderTest.java
@@ -108,8 +108,23 @@ public class DataProviderTest extends ScenarioTest<GivenTestStep, WhenTestStep, 
 
     @Test
     @DataProvider( { "1", "2" } )
-    public void table_parameters_are_ignored_by_the_case_analysis( Integer arg ) {
-        given().a_step_with_a_table_parameter( 1, 2, 3 );
+    public void table_parameters_work_with_primitive_arrays( Integer arg ) {
+        given().a_step_with_a_table_parameter_and_primitive_array( 1, 2, 3 );
+    }
+
+    @Test
+    @DataProvider( { "1", "2" } )
+    public void table_parameters_are_ignored_by_the_case_analysis( Integer arg ) throws Throwable {
+        given().a_step_with_a_table_parameter( new GivenTestStep.TestTableEntry() );
+
+        getScenario().finished();
+        ScenarioModel scenarioModel = getScenario().getModel().getLastScenarioModel();
+        if( scenarioModel.getScenarioCases().size() == 2 ) {
+            CaseArgumentAnalyser analyser = new CaseArgumentAnalyser();
+            analyser.analyze( scenarioModel );
+            assertThat( scenarioModel.getDerivedParameters() ).isEmpty();
+        }
+
     }
 
     @Test