+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Fixed Issues
 
 * Fixed issue with primitive arrays and the @Table annotation [#162](https://github.com/TNG/JGiven/issues/162)
+* Fixed an issue when using the `@Table` parameter that could lead to unwanted parameters in the data table [#161](https://github.com/TNG/JGiven/issues/161)
 
 # v0.9.4